test e fix vari User

+++ src/main/java/it/javaWS/javaws/services/UserService.java
@@ -3,6 +3,7 @@ package it.javaWS.javaws.services;
 import it.javaWS.javaws.models.User;
 import it.javaWS.javaws.repositories.UserRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -10,29 +11,41 @@ import java.util.Optional;
 @Service
 public class UserService {
 
-    private final UserRepository userRepository;
+	private final UserRepository userRepository;
 
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
+	public UserService(UserRepository userRepository) {
+		this.userRepository = userRepository;
+	}
 
-    public User createUser(User user) {
-        return userRepository.save(user);
-    }
+	@Transactional
+	public User createUser(User user) {
 
-    public Optional<User> getUser(Long id) {
-        return userRepository.findById(id);
-    }
+		if (userRepository.findByEmailOrUsername(user.getEmail(), user.getUsername()).size() > 0)
+			return null; // Username o Email già utilizzati
+		return userRepository.save(user);
+	}
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
-    }
-    
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
-    }
-    
-    public void updateUser(User user) {
-    	userRepository.save(user);
-    }
+	@Transactional
+	public Optional<User> getUser(Long id) {
+		return userRepository.findById(id);
+	}
+
+	@Transactional
+	public List<User> getAllUsers() {
+		return userRepository.findAll();
+	}
+
+	@Transactional
+	public Boolean deleteUser(Long id) {
+		if (!getUser(id).isPresent())
+			return false;
+		userRepository.deleteById(id);
+		return true;
+	}
+
+	@Transactional
+	public User updateUser(User user) {
+		return userRepository.save(user);
+	}
+
 }
